chore: initial move/reanme/delete of comopnents
